requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.11
python-decouple==3.8
drf-yasg==1.21.7
gunicorn==21.2.0

